Version 2, implemented: - State Machine Logic - state transition 99% done - communication logic to MBED - hanshaking - servo driver - two-mode operation (maintenance and operation) - using Quartus II Version 13, Service Pack 1 - latest to support Cyclone 2 family of boards
TO FIX:
Getting error on compilation at instantiation of Instruction instr (...) "...must be connected to a structural net expression".

PROPOSED SOLUTION:
servoInstr port of the Instruction module is an output port.
You should not connect an output to a reg in an upper module.
Two outputs should not drive the same wire.